+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Intl API is mandatory now and will not fallback anymore
 - possible compatibility layer for older formats: `test/compatibility/v4/v4Compatibility.js`
 - rename `initImmediate` to `initAsync`
+- fallback to `dev` language if plural rule not found
 - addresses
   - [2244](https://github.com/i18next/i18next/pull/2244)
   - [2184](https://github.com/i18next/i18next/pull/2184)
@@ -17,7 +18,6 @@
   - [2208](https://github.com/i18next/i18next/issues/2208)
 
 ➡️ check out the [migration guide](https://www.i18next.com/misc/migration-guide#v23.x.x-to-v24.0.0)
-
 
 ## 23.16.7